Fix PS1 usage and make it more readable
bash doesn't know when chars aren't displayed via PS1. Thus within PS1
\[ and \] should be used around characters that aren't displayed so they
aren't counted. These are special cases for PS1 specifically so they
need to be inside PS1 when it's value is used. NOT as apart of
expansion/execution within the PS1 otherwise they just come through as
regular characters and don't do the job they're meant to do.

To deal with this I've given PROMPT_COMMAND a value to set the PS1 ahead
of time. This makes it so that PS1 is set to the expanded output BEFORE
it's read for use. Note, PROMPT_COMMAND can be an array of commands if
more comands are desired later.